requirements.txt
tensorflow
keras
Pillow
numpy
tqdm
matplotlib
streamlit
seaborn